+++ src/context/ThemeContext/themeContext.ts
@@ -1,0 +1,19 @@
+import { createContext, Dispatch, SetStateAction, useContext } from "react";
+import { tokensWithReference } from "@src/design/tokensWithReference";
+
+type ThemeName = keyof typeof tokensWithReference;
+
+interface ThemeContextType {
+  themeName: ThemeName;
+  setThemeName: Dispatch<SetStateAction<ThemeName>>;
+}
+
+const ThemeContext = createContext<ThemeContextType>({
+  themeName: "sistemasenlinea",
+  setThemeName: () => {},
+});
+
+const useTheme = () => useContext(ThemeContext);
+
+export { ThemeContext, useTheme };
+export type { ThemeName, ThemeContextType };
